Implement support for cubemap rendering (#671)
- Implement cubemap rendering support into backend
- Implement point shadow rendering demo that uses the cubemap
- Fix inaccurate App timer
- Check for OpenGL extensions at runtime and warn users (might be removed in future)

+++ src/oscar/Platform/AppClock.hpp
@@ -7,7 +7,7 @@ namespace osc
 {
     // definition of application's runtime clock
     struct AppClock final {
-        using rep = float;
+        using rep = double;
         using period = std::ratio<1>;
         using duration = std::chrono::duration<rep, period>;
         using time_point = std::chrono::time_point<AppClock>;
